:wrench:move admin controller
